\name{basetools}
%No index entry without an alias. Bug or feature?
\alias{basetools}
\title{Demonstration Package with Basic Tools}
\description{
  General information about the custom basetools package.
}
\details{
    The package basetools is a simple demonstration package
    for basic tools to build the holiday server example
    main program. It contains basic tools lile strip that
    I did not find in the first place packaged with R.
    However, I did not search for long.

    See \code{library(help=basetools)} for a list of all available functions.

}
\author{Thorsten Rangwich}
\seealso{
  \code{\link{readConfigFile}}
}
