% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_csv.R
\name{sample_csv}
\alias{sample_csv}
\title{Sample lines from csv}
\usage{
sample_csv(file, prob = 1/1000, header = T)
}
\arguments{
\item{file}{a csv file to sample}

\item{prob}{probability of reading a line, defaults to 1/1000}

\item{header}{does file have a header? Defaults to TRUE}
}
\value{
a data frame
}
\description{
Sample lines from csv
}
\keyword{lines}
\keyword{sample}
